Fix broken FileUtils' getAssetsPath()
getAssetsPath() method from neon::utility had a hardcoded requirement
that the build directory is called "bin". This is no longer the case,
and anyway the code should not be dependant on it.

This commit fixes the above problem by rewriting the FileUtils
implementation. In place of readlink and <unistd.h>, std::filesystem
features are being used. Also instead of returning a boolean and a
string, an optional<string> is now being returned.

With these modifications runtime now successfully detects the assets'
directory.